Refs #RHIROS-786 - extract decorator in separate file
